requirements.txt
sq~=1.0a1
marshmallow~=3.20.2
Flask~=3.0.2
pytz~=2024.1
pip~=24.0
pathlib~=1.0.1
configparser~=6.0.1
wheel~=0.42.0
Jinja2~=3.1.3
filelock~=3.13.1
networkx~=3.2.1
nipype~=1.8.6
fitz~=0.0.1.dev2
numpy~=1.26.4
parso~=0.8.3
jedi~=0.19.1
colorama~=0.4.6
lxml~=5.1.0
prov~=2.0.0
rdflib~=7.0.0
pydot~=2.0.0
python-dateutil~=2.9.0
typing_extensions~=4.10.0
MarkupSafe~=2.1.5
Werkzeug~=3.0.1
click~=8.1.7
blinker~=1.7.0
itsdangerous~=2.1.2
setuptools~=69.1.1
pyparsing~=3.1.1
scipy~=1.12.0
packaging~=23.2
ipython~=8.22.1
nibabel~=5.2.1
looseversion~=1.3.0
simplejson~=3.19.2
pandas~=2.2.1
traits~=6.4.3
SQLAlchemy~=2.0.27
requests~=2.31.0
pyxnat~=1.6.2
isodate~=0.6.1
certifi~=2024.2.2
Pygments~=2.17.2
traitlets~=5.14.1
decorator~=5.1.1
wcwidth~=0.2.13
six~=1.16.0
urllib3~=2.2.1
python-dotenv~=1.0.1
greenlet~=3.0.3
httplib2~=0.22.0
idna~=3.6
configobj~=5.0.8
asttokens~=2.4.1
executing~=2.0.1